requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
fastapi==0.112.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
yarl==1.9.4